+++ snekql/mariadb/__init__.py
@@ -137,7 +137,7 @@ else:
     Transaction = _Transaction
     Write = _RuntimeWrite
 
-from snekql.mariadb.recursive import recursive_cte
+from snekql.mariadb.recursive import Cte, NamedOperand, recursive_cte
 
 __all__ = [
     "PENDING_GENERATION",
@@ -154,6 +154,7 @@ __all__ = [
     "CommitOutcome",
     "CompiledQuery",
     "Config",
+    "Cte",
     "CurrentTimestamp",
     "Database",
     "DatabaseCloseTimeoutError",
@@ -199,6 +200,7 @@ __all__ = [
     "ModelError",
     "ModelValidationError",
     "MultipleResultsError",
+    "NamedOperand",
     "NoResultError",
     "Observer",
     "OrderBy",